Update functions to operate on TEXT and remove URI type
In the original code the functions take an object of the URI type as
parameter, and the functions used to convert TEXT to URI unconditionally
abort the query with an error if the uri fails to parse.

I needed functions that operate on strings that may or may not be valid
URIs, with the functions returning NULL if a string fails to parse.